+++ android/app/src/test/java/io/coursepick/coursepick/domain/course/DistanceTest.kt
@@ -1,6 +1,5 @@
-package io.coursepick.coursepick.domain
+package io.coursepick.coursepick.domain.course
 
-import io.coursepick.coursepick.domain.course.Distance
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
