+++ android/app/src/test/java/io/coursepick/coursepick/domain/course/LatitudeTest.kt
@@ -1,6 +1,5 @@
-package io.coursepick.coursepick.domain
+package io.coursepick.coursepick.domain.course
 
-import io.coursepick.coursepick.domain.course.Latitude
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
