+++ android/app/src/test/java/io/coursepick/coursepick/presentation/course/CoursesViewModelTest.kt
@@ -1,13 +1,15 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.course
 
 import io.coursepick.coursepick.domain.course.Course
 import io.coursepick.coursepick.domain.fixture.COORDINATE_FIXTURE
 import io.coursepick.coursepick.domain.fixture.COURSE_FIXTURE_20
 import io.coursepick.coursepick.domain.fixture.FAKE_COURSES
-import io.coursepick.coursepick.presentation.course.CourseItem
+import io.coursepick.coursepick.presentation.extension.CoroutinesTestExtension
+import io.coursepick.coursepick.presentation.extension.InstantTaskExecutorExtension
 import io.coursepick.coursepick.presentation.fixtures.FakeRepository
+import io.coursepick.coursepick.presentation.ui.getOrAwaitValue
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -15,13 +17,13 @@ import org.junit.jupiter.api.extension.ExtendWith
 @ExperimentalCoroutinesApi
 @ExtendWith(CoroutinesTestExtension::class)
 @ExtendWith(InstantTaskExecutorExtension::class)
-class MainViewModelTest {
+class CoursesViewModelTest {
     private val fakeRepository = FakeRepository()
-    private lateinit var mainViewModel: MainViewModel
+    private lateinit var mainViewModel: CoursesViewModel
 
     @BeforeEach
     fun setUp() {
-        mainViewModel = MainViewModel(fakeRepository)
+        mainViewModel = CoursesViewModel(fakeRepository)
         mainViewModel.fetchCourses(COORDINATE_FIXTURE)
     }
 
@@ -29,7 +31,7 @@ class MainViewModelTest {
     fun `초기 상태에서는 가장 가까운 코스가 선택된다`() {
         // given
         val expected =
-            MainUiState(
+            CoursesUiState(
                 FAKE_COURSES.mapIndexed { index: Int, course: Course ->
                     CourseItem(course, selected = index == 0)
                 },
@@ -37,14 +39,14 @@ class MainViewModelTest {
         val actual = mainViewModel.state.getOrAwaitValue()
 
         // then
-        assertThat(actual).isEqualTo(expected)
+        Assertions.assertThat(actual).isEqualTo(expected)
     }
 
     @Test
     fun `하나의 코스를 선택하면 나머지는 해제되고 해당 코스만 선택된다`() {
         // given
         val expected =
-            MainUiState(
+            CoursesUiState(
                 FAKE_COURSES.map { course: Course ->
                     CourseItem(course, selected = course == COURSE_FIXTURE_20)
                 },
@@ -54,7 +56,7 @@ class MainViewModelTest {
         mainViewModel.select(CourseItem(COURSE_FIXTURE_20, selected = false))
 
         // then
-        assertThat(mainViewModel.state.getOrAwaitValue()).isEqualTo(expected)
+        Assertions.assertThat(mainViewModel.state.getOrAwaitValue()).isEqualTo(expected)
     }
 
     @Test
@@ -63,7 +65,7 @@ class MainViewModelTest {
         mainViewModel.select(CourseItem(COURSE_FIXTURE_20, selected = false))
 
         val expected =
-            MainUiState(
+            CoursesUiState(
                 FAKE_COURSES.map { course: Course ->
                     CourseItem(course, selected = course == COURSE_FIXTURE_20)
                 },
@@ -73,6 +75,6 @@ class MainViewModelTest {
         mainViewModel.select(CourseItem(COURSE_FIXTURE_20, selected = true))
 
         // then
-        assertThat(mainViewModel.state.getOrAwaitValue()).isEqualTo(expected)
+        Assertions.assertThat(mainViewModel.state.getOrAwaitValue()).isEqualTo(expected)
     }
 }
