+++ android/app/src/test/java/io/coursepick/coursepick/presentation/extension/CoroutinesTestExtension.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.extension
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
