+++ android/app/src/test/java/io/coursepick/coursepick/presentation/extension/InstantTaskExecutorExtension.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.extension
 
 import androidx.arch.core.executor.ArchTaskExecutor
 import androidx.arch.core.executor.TaskExecutor
